Peacockfan

+++ Test/PlayerLogicComponent.cpp
@@ -141,7 +141,7 @@ void PlayerLogicComponent::addToKillList(uint64_t shooter, uint64_t victim){
 	//start by finding the first open spot to add the notification to
 	int elemIndex = 0;
 	bool success = false;
-	for (int i = 0; i < killHUD.size(); i++){
+	for (unsigned i = 0; i < killHUD.size(); i++){
 		if (!killHUD[i].first->visible){
 			elemIndex = i;
 			success = true;
@@ -153,7 +153,7 @@ void PlayerLogicComponent::addToKillList(uint64_t shooter, uint64_t victim){
 		//this happens when we iterate through the list and see that there is no free spot to put the new notification
 		//we need to rotate the array and add it on to the end now
 		auto tempElem = killHUD[0];
-		for (int i = 0; i < killHUD.size() - 1; i++){
+		for (unsigned i = 0; i < killHUD.size() - 1; i++){
 			killHUD[i] = killHUD[i + 1];
 		}
 		killHUD[killHUD.size() - 1] = tempElem;
@@ -179,13 +179,13 @@ void PlayerLogicComponent::updateKillHUD(){
 		//circular rotation on array
 		killHUD[0].first->visible = false;
 		auto tempElem = killHUD[0];
-		for (int i = 0; i < killHUD.size() - 1; i++){
+		for (unsigned i = 0; i < killHUD.size() - 1; i++){
 			killHUD[i] = killHUD[i+1];
 		}
 		killHUD[killHUD.size() - 1] = tempElem;
 		//now update postions on screen
-		for (int i = 0; i < killHUD.size(); i++){
-			killHUD[i].first->setPos(SCREEN_WIDTH - 200, 130 + i * 30);
+		for (unsigned i = 0; i < killHUD.size(); i++){
+			killHUD[i].first->setPos((float)(SCREEN_WIDTH - 200), (float)(130 + i * 30));
 		}
 	}
 }
